requirements.txt
requests
numpy
matplotlib
numpy-stl